refactor: DatabaseCleaner에서 table 이름 조회 로직을 바꾸어 Like 엔티티 이름 원래대로 사용하도록 변경
